+++ modules/utils.py
@@ -13,13 +13,12 @@ import yaml
 DIR_PATH = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 NAGIOS_PATH = "/usr/lib/nagios/plugins/"
 
-# Redis Keys
-VALID_HOSTS = "host_names"
-HOST_STATUS = "host_status"
-COMMAND_TASK_ID = "command_task_id"
-
-# status description
+# valid status descriptions
 SERVICE_STATUSES = ["OK", "Warning", "Critical", "Unknown"]
+
+# valid service state statuses
+CONFIRMED_STATE = "CONFIRMED"
+UNCONFIRMED_STATE = "UNCONFIRMED"
 
 # allowed file types for web editor
 ALLOWED_EDITOR_TYPES = ('.yaml', '.py')
@@ -54,21 +53,6 @@ def read_yaml(file):
         logging.error(f"Error parsing YAML file {file}")
 
     return result
-
-
-# read a key from the database, converting to dict
-def read_db(db, db_key):
-    result = {}
-
-    if(db.exists(db_key)):
-        result = json.loads(db.get(db_key))
-
-    return result
-
-
-# write a value to the datase, converting to JSON string
-def write_db(db, db_key, db_value):
-    db.set(db_key, json.dumps(db_value))
 
 
 # write JSON to file